requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
supabase==2.16.0